chore(filesystem): move routes to pages directory
