+++ app/src/pages/Edit.tsx
@@ -2,16 +2,12 @@ import Editor, { OnMount } from "@monaco-editor/react";
 import { useThrottleCallback } from "@react-hook/throttle";
 import merge from "deepmerge";
 import { stringify } from "gray-matter";
-import { Resizable } from "re-resizable";
-import {
-  Suspense,
-  useCallback,
-  useEffect,
-  useReducer,
-  useRef,
-  useState,
-} from "react";
+import { useCallback, useEffect, useReducer, useRef, useState } from "react";
 
+import EditorError from "../components/EditorError";
+import GraphProvider from "../components/GraphProvider";
+import Loading from "../components/Loading";
+import useGraphOptions from "../components/useGraphOptions";
 import {
   delimiters,
   editorOptions,
@@ -25,23 +21,17 @@ import {
   themeNameDark,
   themeNameLight,
 } from "../lib/registerLanguage";
-import Docs from "./Docs";
 import styles from "./Edit.module.css";
-import EditorError from "./EditorError";
-import GraphProvider from "./GraphProvider";
-import helpStyles from "./Help.module.css";
-import Loading from "./Loading";
-import useGraphOptions from "./useGraphOptions";
 
-export default function Help() {
-  const [text, setText] = useLocalStorageText("h"); // fixed workspace name
+export default function Edit() {
+  const [text, setText] = useLocalStorageText();
   const [textToParse, setTextToParse] = useReducer(
     (t: string, u: string) => u,
     text
   );
   const setTextToParseThrottle = useThrottleCallback(setTextToParse, 2, true);
   const [hoverLineNumber, setHoverLineNumber] = useState<undefined | number>();
-  const editorRef = useRef<null | Parameters<OnMount>[0]>(null);
+  const editorRef = useRef<Parameters<OnMount>[0]>(null);
   const monacoRef = useRef<any>();
   const { data: mode } = useAppMode();
   const loading = useRef(<Loading />);
@@ -78,13 +68,6 @@ export default function Help() {
   // Hover
   useEditorHover(editorRef, hoverLineNumber);
 
-  useEffect(() => {
-    window.flowchartFunSetHelpText = setText;
-    return () => {
-      delete window.flowchartFunSetHelpText;
-    };
-  }, [setText]);
-
   const onChange = useCallback((value) => setText(value ?? ""), [setText]);
 
   return (
@@ -95,46 +78,16 @@ export default function Help() {
       graphOptions={graphOptions}
       updateGraphOptionsText={updateGraphOptionsText}
     >
-      <div className={helpStyles.helpWrapper} data-testid="help">
-        <Resizable
-          defaultSize={{ width: "100%", height: "50vh" }}
-          className={helpStyles.resizable}
-          enable={{
-            top: false,
-            right: false,
-            bottom: true,
-            left: false,
-            topRight: false,
-            bottomRight: false,
-            bottomLeft: false,
-            topLeft: false,
-          }}
-        >
-          <div className={helpStyles.docsWrapper}>
-            <div className={helpStyles.docsWrapperScroll}>
-              <Suspense fallback={<Loading />}>
-                <Docs currentText={text} />
-              </Suspense>
-            </div>
-          </div>
-        </Resizable>
-        <Editor
-          value={text}
-          wrapperClassName={styles.Editor}
-          defaultLanguage={languageId}
-          options={editorOptions}
-          onChange={onChange}
-          loading={loading.current}
-          onMount={onMount}
-        />
-      </div>
+      <Editor
+        value={text}
+        wrapperClassName={styles.Editor}
+        defaultLanguage={languageId}
+        options={editorOptions}
+        onChange={onChange}
+        loading={loading.current}
+        onMount={onMount}
+      />
       <EditorError />
     </GraphProvider>
   );
-}
-
-declare global {
-  interface Window {
-    flowchartFunSetHelpText?: (text: string) => void;
-  }
 }
