+++ app/src/pages/Help.tsx
@@ -2,8 +2,21 @@ import Editor, { OnMount } from "@monaco-editor/react";
 import { useThrottleCallback } from "@react-hook/throttle";
 import merge from "deepmerge";
 import { stringify } from "gray-matter";
-import { useCallback, useEffect, useReducer, useRef, useState } from "react";
+import { Resizable } from "re-resizable";
+import {
+  Suspense,
+  useCallback,
+  useEffect,
+  useReducer,
+  useRef,
+  useState,
+} from "react";
 
+import Docs from "../components/Docs";
+import EditorError from "../components/EditorError";
+import GraphProvider from "../components/GraphProvider";
+import Loading from "../components/Loading";
+import useGraphOptions from "../components/useGraphOptions";
 import {
   delimiters,
   editorOptions,
@@ -18,20 +31,17 @@ import {
   themeNameLight,
 } from "../lib/registerLanguage";
 import styles from "./Edit.module.css";
-import EditorError from "./EditorError";
-import GraphProvider from "./GraphProvider";
-import Loading from "./Loading";
-import useGraphOptions from "./useGraphOptions";
+import helpStyles from "./Help.module.css";
 
-export default function Edit() {
-  const [text, setText] = useLocalStorageText();
+export default function Help() {
+  const [text, setText] = useLocalStorageText("h"); // fixed workspace name
   const [textToParse, setTextToParse] = useReducer(
     (t: string, u: string) => u,
     text
   );
   const setTextToParseThrottle = useThrottleCallback(setTextToParse, 2, true);
   const [hoverLineNumber, setHoverLineNumber] = useState<undefined | number>();
-  const editorRef = useRef<Parameters<OnMount>[0]>(null);
+  const editorRef = useRef<null | Parameters<OnMount>[0]>(null);
   const monacoRef = useRef<any>();
   const { data: mode } = useAppMode();
   const loading = useRef(<Loading />);
@@ -68,6 +78,13 @@ export default function Edit() {
   // Hover
   useEditorHover(editorRef, hoverLineNumber);
 
+  useEffect(() => {
+    window.flowchartFunSetHelpText = setText;
+    return () => {
+      delete window.flowchartFunSetHelpText;
+    };
+  }, [setText]);
+
   const onChange = useCallback((value) => setText(value ?? ""), [setText]);
 
   return (
@@ -78,16 +95,46 @@ export default function Edit() {
       graphOptions={graphOptions}
       updateGraphOptionsText={updateGraphOptionsText}
     >
-      <Editor
-        value={text}
-        wrapperClassName={styles.Editor}
-        defaultLanguage={languageId}
-        options={editorOptions}
-        onChange={onChange}
-        loading={loading.current}
-        onMount={onMount}
-      />
+      <div className={helpStyles.helpWrapper} data-testid="help">
+        <Resizable
+          defaultSize={{ width: "100%", height: "50vh" }}
+          className={helpStyles.resizable}
+          enable={{
+            top: false,
+            right: false,
+            bottom: true,
+            left: false,
+            topRight: false,
+            bottomRight: false,
+            bottomLeft: false,
+            topLeft: false,
+          }}
+        >
+          <div className={helpStyles.docsWrapper}>
+            <div className={helpStyles.docsWrapperScroll}>
+              <Suspense fallback={<Loading />}>
+                <Docs currentText={text} />
+              </Suspense>
+            </div>
+          </div>
+        </Resizable>
+        <Editor
+          value={text}
+          wrapperClassName={styles.Editor}
+          defaultLanguage={languageId}
+          options={editorOptions}
+          onChange={onChange}
+          loading={loading.current}
+          onMount={onMount}
+        />
+      </div>
       <EditorError />
     </GraphProvider>
   );
+}
+
+declare global {
+  interface Window {
+    flowchartFunSetHelpText?: (text: string) => void;
+  }
 }
