+++ app/src/pages/ReadOnly.tsx
@@ -3,6 +3,9 @@ import strip from "@tone-row/strip-comments";
 import matter from "gray-matter";
 import { useEffect, useRef, useState } from "react";
 
+import EditorError from "../components/EditorError";
+import GraphProvider from "../components/GraphProvider";
+import Loading from "../components/Loading";
 import { delimiters, editorOptions } from "../lib/constants";
 import { useEditorHover, useEditorOnMount } from "../lib/editorHooks";
 import { useReadOnlyText } from "../lib/hooks";
@@ -12,9 +15,6 @@ import {
   themeNameDark,
   themeNameLight,
 } from "../lib/registerLanguage";
-import EditorError from "./EditorError";
-import GraphProvider from "./GraphProvider";
-import Loading from "./Loading";
 import styles from "./ReadOnly.module.css";
 
 function ReadOnly() {
